Check the cloud sources nickname before change
resolves #10

This checks for `Created by Sources` nickname before altering the
nickname for the cloud access record.

Cloud Sources[1] has an automated process that updates a record in Cloud
Access[2]. When this nickname is changed, Cloud Sources will
periodically update the Cloud Access record and overwrite and nickname
that was altered.

This results in this role / module and Cloud Sources overwriting each
other's changes everytime this role / module is ran.

1: https://console.redhat.com/settings/sources
2: https://access.redhat.com/management/cloud