Updated detect_provider to account for AWS changes
AWS does not put `amazon` in the bios version. Updated the task to check
in multiple ansible facts for amazon:

* ansible_bios_version
* ansible_bios_vendor
* ansible_chassis_asset_tag
* ansible_chassis_vendor

Also resolves #9 - GCE command was not using become, and was returning
an error if the grep string was not found. Summary of changes where:

* changed task from script to command
* removed grep
* added become instead of sudo command
* removed fact creation task
* changed GCE string check to be against raw stdout

Added checks in case bios and board facts missing

Just in case if all cloud providers, VMs, or physical systems will not
define all the ansible facts that are being used to detect the AWS
provider, changed the way the variable is checked by ensuring missing
facts do not raise an error.